+++ pg-core/src/java/org/pg/Connection.java
@@ -1416,15 +1416,15 @@ public final class Connection implements AutoCloseable {
     @SuppressWarnings("unused")
     public int pollNotifications() {
         try (TryLock ignored = lock.get()) {
-            final Result res = new Result(ExecuteParams.INSTANCE, "");
-            while ( IOTool.available(inStream) > 0 ) {
+            final Result res = new Result(ExecuteParams.INSTANCE, "--pollNotifications");
+            while (IOTool.available(inStream) > 0) {
                 final IServerMessage msg = readMessage(res.hasException());
-                if ( msg instanceof NotificationResponse ||
-                     msg instanceof NoticeResponse ||
-                     msg instanceof ParameterStatus ) {
+                if (   msg instanceof NotificationResponse
+                    || msg instanceof NoticeResponse
+                    || msg instanceof ParameterStatus) {
                     handleMessage(msg, res);
                 } else {
-                    throw new PGError("Unexpected message: %s", msg);
+                    throw new PGError("Unexpected message in pollNotifications: %s", msg);
                 }
             }
             res.maybeThrowError();