+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## 0.1.34-SNAPSHOT
 
+- ?
 - #39 Introduce error hierarchy, PGBaseError, PGIOException @jgdavey
 
 ## 0.1.33