+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - RowMap: do not share the Connection's TryLock instance
 - RowMap: implement Indexed and IDeref interfaces
 - #50 Use Executor rather than ExecutorService @jgdavey
+- Refactor CodecParams class (make it mutable)
 
 ## 0.1.33